created add pet route on front end

+++ public/js/pet.js
@@ -2,7 +2,6 @@ async function petAddForm(event) {
   event.preventDefault();
   const pet_name = document.querySelector("#added-pet").value;
 
-  // Send fetch request to add a new dish
   const response = await fetch("/api/pet", {
     method: "POST",
     body: JSON.stringify({
@@ -12,7 +11,6 @@ async function petAddForm(event) {
       "Content-Type": "application/json",
     },
   });
-  // if the dish is added, the 'all' template will be rerendered
   if (response.ok) {
     document.location.replace("/profile");
   } else {
@@ -20,40 +18,21 @@ async function petAddForm(event) {
   }
 }
 
-async function removePet(event) {
-  event.preventDefault();
-  const pet_name = document.querySelector("#added-pet").value;
+const delButtonHandler = async (event) => {
+  if (event.target.hasAttribute("data-id")) {
+    const id = event.target.getAttribute("data-id");
 
-  // Send fetch request to add a new dish
-  const response = await fetch("/api/pet/:id", {
-    method: "DELETE",
-  });
-  // if the dish is added, the 'all' template will be rerendered
-  if (response.ok) {
-    document.location.replace("/profile");
-  } else {
-    alert("Failed to remove a pet");
+    const response = await fetch(`/api/pet/${id}`, {
+      method: "DELETE",
+    });
+
+    if (response.ok) {
+      document.location.replace("/profile");
+    } else {
+      alert("Failed to delete pet");
+    }
   }
-}
+};
 
 document.querySelector("#petAdd").addEventListener("click", petAddForm);
-document.querySelector("#petAdd").addEventListener("click", petAddForm);
-
-// const delButtonHandler = async (event) => {
-//   if (event.target.hasAttribute("data-id")) {
-//     const id = event.target.getAttribute("data-id");
-
-//     const response = await fetch(`/api/projects/${id}`, {
-//       method: "DELETE",
-//     });
-
-//     if (response.ok) {
-//       document.location.replace("/profile");
-//     } else {
-//       alert("Failed to delete project");
-//     }
-//   }
-// };
-// document
-//   .querySelector(".project-list")
-//   .addEventListener("click", delButtonHandler);
+document.querySelector(".pet-list").addEventListener("click", delButtonHandler);
